+++ eslint.config.js
@@ -38,7 +38,7 @@ export default tseslint.config({
   rules: {
     "@typescript-eslint/no-explicit-any": "off",
     "no-unused-vars": "off",
-    "@typescript-eslint/no-unused-vars": ["warn"],
+    "@typescript-eslint/no-unused-vars": "off",
     "no-console": "off",
     "workspaces/no-relative-imports": "off",
     "workspaces/require-dependency": "off",